Reifying Pathfinders schemas for extraction, processing and validation
For Pathfinders, we use a single schema to configure extraction of 'tables' from spreadsheet worksheets, processing of those tables to create clean, labelled DataFrames, and validation of those cleaned DataFrames using Pandera. Currently, the schema is a plain Python object. This change introduces new classes for representing the objects and sub-objects that form the schema, and uses those new classes in the schema.