FPASF-599 - PF R2 - accounting for textual discrepancies in outputs and outcomes between rounds 1 and 2
A copy and paste of outputs and outcomes from an unverified location into the R2 spreadsheet that was sent out to LAs broke the link between the outputs and outcomes control tables in the hidden worksheets and the outputs and outcomes user tables in the user-facing worksheets. An analysis identified several small discrepancies between the outputs and outcomes as they appear in round 1 and the outputs and outcomes as they appear in round 2, specifically differences in whitespace, capitalisation and in one case, a spelling fix ('Ratinalised').

Because we are going to apply the spelling fix manually to production data, we need to apply the spelling fix as a transform in the PF round 1 ingest pipeline. And then because the R2 spreadsheets have already been sent out to LAs with the old text for the bespoke outputs in the hidden worksheets, we need to make cross-table validation checks case-insensitive and handle the spelling fix case in PF round 2 cross-table validation. We don't need to apply a transform in the PF round 2 ingest pipeline as the text coming in is fine. And we don't need to modify cross-table validation checks for PF round 1 because the hidden worksheets align with the user-facing worksheets.

Please note also that there was no need to handle whitespace stripping, as this is already done during table extraction.