Re-skin the admin action pages
A bit of threading-the-needle required here to get classes onto the form
input fields.

`lib.render_form` is part of flask-admin's form rendering pipeline,
which takes a `form_opts` parameter (normally a FormOpts instance). This
has an attribute `widget_args` holding a dict of field names to widget
args.

The widget args we specify here get passed to the govuk-frontend-wtf
widgets, which take a `params` argument that broadly match GOV.UK
Frontend's `params` for nunjucks.