FSPT-190: get account store uri from the parameter store
Fetching the account store uri from the parameter store will allow us to
switch from using the existing fixed account store microservice and
point it to the new combined pre-award-stores when its ready.

We'll need to update the workflows repository which seems to define
related lambda configuration here but i'll look to do that separately.