FLS-1132 - Towns Fund Round 7 - Initial fixes to support upload if initial validation is disabled
The only difference between this round's spreadsheet and the last (other than "Reporting Period" date range change) is that in each of the tables in the "4a - Funding Profiles" tab, th
e name of the penultimate column has been changed from "Beyond 25/26" to "Financial Year 2026/27 (£s)". We need to make sure we map this new column name to a start and end date in the "convert_financial_halves" utility function, while continuing to support "Beyond 25/26", because one of the columns in the outputs tables maps its name to this.