FLS-1132 - Towns Fund Round 7 - Prepare new round to open conditional on new env var ENABLE_TF_R7
This won't actually open the round, it will just mean that on the day we want the round to open, we can go into AWS and switch ENABLE_TF_R7 from False to True, with no need for deployment, in line with previous round openings. This change will however CLOSE the previous round.