Pathfinders R3 actual / forecast column name fix
For Pathfinders R3, when Bolton Council tried to submit their monitoring return, we got an error in Sentry: the columns in the spreadsheet did not match the columns specified in the backend schema. Specifically, in the spreadsheet we use 'Actual' instead of 'Forecast' for FY 24-25 Oct-Dec and Jan-Mar, while we have 'Forecast' for these periods in the backend schema. The spreadsheet is correct here, so we should update the backend schema accordingly.